requirements.txt
jsonschema
pytest-allure-adaptor
requests